Aplicando o banco de dados em todas as views

+++ src/models/Product.js
@@ -9,7 +9,7 @@ module.exports = (sequelize, DataTypes) => {
             allowNull: false
         },
         price: {
-            type: DataTypes.DOUBLE,
+            type: DataTypes.DECIMAL,
             allowNull: false
         },
         image: {
